FoodPlanRepo done (#68)
* FoodPlanRepo done
* sql fixed for select
* update, mapRow & sql fixed in FoodPlanRepo

+++ domain-tier/src/main/java/com/outdoor/foodcalc/domain/service/plan/FoodPlanDomainService.java
@@ -1,12 +1,12 @@
 package com.outdoor.foodcalc.domain.service.plan;
 
+import com.outdoor.foodcalc.domain.exception.FoodcalcDomainException;
 import com.outdoor.foodcalc.domain.exception.NotFoundException;
 import com.outdoor.foodcalc.domain.model.plan.FoodPlan;
 import com.outdoor.foodcalc.domain.repository.plan.IFoodPlanRepo;
 import com.outdoor.foodcalc.domain.service.FoodPlansRepo;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,29 +22,43 @@ public class FoodPlanDomainService {
     }
 
     public List<FoodPlan> getAllFoodPlans() {
-        return new ArrayList<>(tmpRepo.getAllPlans());
+        List<FoodPlan> plans = planRepo.getAllFoodPlans();
+        plans.forEach(plan -> plan.setDays(tmpRepo.getPlanDays(plan.getId())));
+        return plans;
     }
 
     public Optional<FoodPlan> getFoodPlan(long id) {
-        return Optional.ofNullable(tmpRepo.getFoodPlan(id));
-    }
-
-    public void deleteFoodPlan(long id) {
-        tmpRepo.deleteFoodPlan(id);
+        return planRepo.getFoodPlan(id).map(plan -> {
+           plan.setDays(tmpRepo.getPlanDays(plan.getId()));
+           return plan;
+        });
     }
 
     public FoodPlan addFoodPlan(FoodPlan plan) {
+        long id = planRepo.addFoodPlan(plan);
         FoodPlan newPlan = plan.toBuilder()
-                .id(tmpRepo.getMaxPlanIdAndIncrement())
+                .id(id)
                 .build();
-        tmpRepo.addFoodPlan(newPlan);
+        tmpRepo.addFoodPlan(id);
         return newPlan;
     }
 
     public void updateFoodPlan(FoodPlan plan) {
-        if(tmpRepo.getFoodPlan(plan.getId()) == null) {
-            throw new NotFoundException("FoodPlan with id=" + plan.getId() + " doesn't exist");
+        if(!planRepo.existsFoodPlan(plan.getId())) {
+            throw new NotFoundException("Food plan with id=" + plan.getId() + " doesn't exist");
         }
-        tmpRepo.updateFoodPlan(plan);
+        if(!planRepo.updateFoodPlan(plan)) {
+            throw new FoodcalcDomainException("Failed to update product category with id=" + plan.getId());
+        }
+    }
+
+    public void deleteFoodPlan(long id) {
+        if(!planRepo.existsFoodPlan(id)) {
+            throw new NotFoundException("Food plan with id=" + id + " doesn't exist");
+        }
+        if (!planRepo.deleteFoodPlan(id)) {
+            throw new FoodcalcDomainException("Failed to delete Food plan with id=" + id);
+        }
+        tmpRepo.deleteFoodPlan(id);
     }
 }
